002_etfbot/drafts/homework/w4 homework_example1/requirements.txt
gradio>=4.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
rank-bm25>=0.2.2
kiwipiepy>=0.18.0
sentence-transformers>=3.0.0
matplotlib>=3.8.0
hypothesis>=6.0.0
